pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tensor-training-core"
version = "0.1.0"
description = "TensorFlow Lite object detection training core."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Louis Chuang" }]
dependencies = [
  "numpy>=1.26,<3",
  "Pillow>=10,<12",
  "pydantic>=2.7,<3",
  "PyYAML>=6.0,<7",
  "typing-extensions>=4.10,<5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9",
  "ruff>=0.4,<1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
